requirements.txt
boto3==1.20.26
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-JWT-Extended==4.3.1
Flask-SQLAlchemy==2.5.1
flask-marshmallow==0.14.0
GeoAlchemy2==0.10.0
marshmallow-sqlalchemy==0.27.0
Shapely==1.8.0
PyJWT==2.3.0
psycopg2==2.9.1